Extract history_path + status_path into a new paths module (#39)
history_path() in persistence.rs and status_path() in waybar.rs
were two near-identical "join a base dir against a fixed filename"
helpers. Lifting them into a new src/paths.rs module puts the two
runtime-artifact filenames in one place, so #34's eventual
mac-notifications-* -> nwg-notifications-* rename is a single-file
change instead of two-file synchronization.

Both helpers keep their existing fallback semantics: history_path
falls back to /tmp when nwg_common can't resolve a cache dir;
status_path falls back to /tmp when XDG_RUNTIME_DIR isn't set.

The persistence.rs callers (load_history, save_history) take a
&Path parameter, so they didn't need an import update — only the
single persistence::history_path() call site in main.rs moves to
paths::history_path(). status_path's only caller is inside
waybar::update_status, which now reads crate::paths::status_path().

Trimmed two unused PathBuf imports while there (persistence.rs
and waybar.rs both lose their PathBuf import after the helpers
move out).

No behavioral change: cargo build clean, full 87-test suite still
green.

Closes #39.
